requirements.txt
pyyaml
slackclient
